% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_cuboids}
\alias{detect_cuboids}
\title{Detect cuboids (rectangular boxes) in point cloud data}
\usage{
detect_cuboids(
  xyz,
  min_width = 0.5,
  max_width = 10,
  min_height = 0.5,
  max_height = 10,
  min_votes = 150L,
  plane_threshold = 0.1,
  label_points = FALSE
)
}
\arguments{
\item{xyz}{A matrix with 3 columns (X, Y, Z coordinates)}

\item{min_width}{Minimum cuboid width}

\item{max_width}{Maximum cuboid width}

\item{min_height}{Minimum cuboid height}

\item{max_height}{Maximum cuboid height}

\item{min_votes}{Minimum number of points required per cuboid}

\item{plane_threshold}{Distance threshold for plane detection}

\item{label_points}{If TRUE, label all points with cuboid ID}
}
\value{
A list with detected cuboids and optionally labeled points
}
\description{
Detect cuboids (rectangular boxes) in point cloud data
}
