% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_spheres}
\alias{detect_spheres}
\title{Detect spheres in point cloud data using 3D Hough transform}
\usage{
detect_spheres(
  xyz,
  min_radius = 0.1,
  max_radius = 2,
  min_votes = 30L,
  label_points = FALSE
)
}
\arguments{
\item{xyz}{A matrix with 3 columns (X, Y, Z coordinates)}

\item{min_radius}{Minimum sphere radius}

\item{max_radius}{Maximum sphere radius}

\item{min_votes}{Minimum number of points required to detect a sphere}

\item{label_points}{If TRUE, label all points with sphere ID}
}
\value{
A list with detected spheres and optionally labeled points
}
\description{
Detect spheres in point cloud data using 3D Hough transform
}
