% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_tree_boles}
\alias{detect_tree_boles}
\title{Detect tree boles (cylindrical trunks) in point cloud data}
\usage{
detect_tree_boles(
  xyz,
  min_radius = 0.05,
  max_radius = 0.5,
  min_height = 1,
  min_votes = 30L,
  vertical_only = TRUE,
  label_points = FALSE
)
}
\arguments{
\item{xyz}{A matrix with 3 columns (X, Y, Z coordinates)}

\item{min_radius}{Minimum trunk radius in meters}

\item{max_radius}{Maximum trunk radius in meters}

\item{min_height}{Minimum visible trunk height in meters}

\item{min_votes}{Minimum number of points required to detect a tree}

\item{vertical_only}{If TRUE, only detect vertical cylinders}

\item{label_points}{If TRUE, label all points with tree ID}
}
\value{
A list with detected trees and optionally labeled points
}
\description{
Detect tree boles (cylindrical trunks) in point cloud data
}
