\name{houghPrimitives-package}
\alias{houghPrimitives-package}
\alias{houghPrimitives}
\docType{package}
\title{
  houghPrimitives: 3D Shape Detection in Point Clouds Using Hough Transforms
}
\description{
  Detect geometric primitives (lines, planes, spheres, cylinders) in 3D point cloud data using Hough transform algorithms.
}
\details{
  This package provides tools for detecting geometric shapes in 3D point clouds, commonly used in LiDAR and photogrammetry applications.
  
  The main function currently available is \code{summarize_points()}, which provides basic statistics about point cloud data.
}
\author{
Author: Aaron Maxwell

Maintainer: Aaron Maxwell <ajsm@umd.edu>
}
\keyword{ package }
