% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summarize_points}
\alias{summarize_points}
\title{Simple point cloud summary}
\usage{
summarize_points(xyz)
}
\arguments{
\item{xyz}{A matrix with 3 columns (X, Y, Z coordinates)}
}
\value{
A list with summary statistics
}
\description{
Simple point cloud summary
}
